+++ src/hittable.rs
@@ -1,8 +1,6 @@
-use std::fmt;
-
 use crate::hitrecord::HitRecord;
 use crate::ray::Ray;
 
 pub trait Hittable {
-    fn hit(&self, ray: &Ray, tmin: f64, tmax: f64) -> Option<HitRecord>;
+    fn hit(&self, ray: &Ray, tmin: f32, tmax: f32) -> Option<HitRecord>;
 }
