+++ src/item.rs
@@ -7,19 +7,19 @@ use crate::vec3::Vec3;
 #[derive(Clone, Copy, Debug)]
 pub struct Sphere {
     center: Vec3,
-    radius: f64,
+    radius: f32,
     material: Material,
     attenuation: Vec3,
-    metal_fuzz: f64,
+    metal_fuzz: f32,
 }
 
 impl Sphere {
     pub fn new(
         center: Vec3,
-        radius: f64,
+        radius: f32,
         material: Material,
         attenuation: Vec3,
-        metal_fuzz: f64,
+        metal_fuzz: f32,
     ) -> Sphere {
         Sphere {
             center,
@@ -32,7 +32,7 @@ impl Sphere {
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, ray: &Ray, tmin: f64, tmax: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, tmin: f32, tmax: f32) -> Option<HitRecord> {
         let oc = ray.origin() - self.center;
         let a = ray.direction().length_squared();
         let b = oc.dot(ray.direction()) * 2.0;
@@ -45,7 +45,7 @@ impl Hittable for Sphere {
             None
         } else {
             // 解がある場合はtが小さいほう(原点から近いほう)のHitRecordを返す
-            let t: f64 = (-b - discriminant.sqrt()) / (2.0 * a);
+            let t: f32 = (-b - discriminant.sqrt()) / (2.0 * a);
 
             if tmin < t && t < tmax {
                 let point = ray.at(t);
