+++ src/ray.rs
@@ -18,7 +18,7 @@ impl Ray {
         self.origin
     }
 
-    pub fn at(&self, t: f64) -> Vec3 {
+    pub fn at(&self, t: f32) -> Vec3 {
         self.origin + self.direction * t
     }
 }
